rename test_config to testfile; it really is about the file
